First set of changes, all moved into one project, no big refactoring done yet apart from some names
