+++ plugins/apg-ssh/src/main/resources/META-INF/gradle-plugins/com.apgsga.ssh.properties
@@ -1,0 +1,1 @@
+implementation-class=com.apgsga.ssh.plugins.ApgSsh
