--- plugins/rpm-ssh-deployer/src/main/resources/META-INF/gradle-plugins/com.apgsga.rpm.ssh.deployer.properties
@@ -1,1 +1,0 @@
-implementation-class=com.apgsga.gradle.ssh.plugin.ApgRpmSshDeployer
