--- plugins/ssh-common/build.gradle
@@ -1,14 +1,0 @@
-apply from: "${projectDir}/../gradle-common/common-plugin.gradle"
-apply from: "${projectDir}/../gradle-common/integration-test.gradle"
-apply from: "${projectDir}/../gradle-common/functional-test.gradle"
-
-description 'Provides properties required for SSH connection'
-
-dependencies {
-    testCompile gradleTestKit()
-    testCompile project(':common-test')
-    // TODO JHE: common-repo needed here because of our test which by default include ENUM coming from common-repo
-    //           BUT: in worse case, testCompile should be enough, doesn't seem to be the case
-    //                best case, this dependency should be resolved within common-test or within gradle-common (functionalTest.gradle)
-    compile project(':common-repo')
-}
