--- plugins/ssh-common/src/main/resources/META-INF/gradle-plugins/com.apgsga.ssh.common.properties
@@ -1,1 +1,0 @@
-implementation-class=com.apgsga.ssh.common.plugin.ApgSshCommonPlugin
