--- plugins/zip-ssh-deployer/src/main/resources/META-INF/gradle-plugins/com.apgsga.zip.ssh.deployer.properties
@@ -1,1 +1,0 @@
-implementation-class=com.apgsga.gradle.ssh.zip.deployer.plugins.ApgZipSshDeployer
